Fix SsdpClient.ssdp_st/0
Use Conform at hex.pm v1.0.0-rc8 instead of github

On invoking `cell list` I hit an error which conform raises internally
when file is missing.

Instead of returning an {:error, :enoent} as the
ssdp_st/0 expected, it was actually raising.

The upgrade made no difference to this error case, however it changed
the success case slightly, returning a tuple like {:ok, conf}.

I've adapted the function to handle if the file exists or not correctly
in each case without crashing, using the latest conform.